fix(Card): demo拆解与规范
