+++ src/packages/card/doc.zh-TW.md
@@ -16,35 +16,7 @@ import { Card, Price, Tag } from '@nutui/nutui-react'
 
 :::demo
 
-```tsx
-import React from 'react'
-import { Card, Price, Tag } from '@nutui/nutui-react'
-
-const App = () => {
-  const state = {
-    src: '//img10.360buyimg.com/n2/s240x240_jfs/t1/210890/22/4728/163829/6163a590Eb7c6f4b5/6390526d49791cb9.jpg!q70.jpg',
-    title:
-      '【活蟹】湖塘煙雨 陽澄湖大閘蟹公4.5兩 母3.5兩 4對8只 鮮活生鮮螃蟹現貨水產禮盒海鮮水',
-    price: '388',
-    vipPrice: '378',
-    shopDescription: '自營',
-    delivery: '廠商配送',
-    shopName: '陽澄湖大閘蟹自營店>',
-  }
-  return (
-    <Card
-      src={state.src}
-      title={state.title}
-      price={state.price}
-      vipPrice={state.vipPrice}
-      shopDescription={state.shopDescription}
-      delivery={state.delivery}
-      shopName={state.shopName}
-     />
-  )
-}
-export default App
-```
+<CodeBlock src='h5/demo1.tsx'></CodeBlock>
 
 :::
 
@@ -52,55 +24,7 @@ export default App
 
 :::demo
 
-```tsx
-import React from "react";
-import { Card, Price, Tag } from '@nutui/nutui-react';
-
-const App = () => {
-  const state = {
-    src:
-      '//img10.360buyimg.com/n2/s240x240_jfs/t1/210890/22/4728/163829/6163a590Eb7c6f4b5/6390526d49791cb9.jpg!q70.jpg',
-    title: '【活蟹】湖塘煙雨 陽澄湖大閘蟹公4.5兩 母3.5兩 4對8只 鮮活生鮮螃蟹現貨水產禮盒海鮮水',
-    price: '388',
-    vipPrice: '378',
-    shopDescription: '自營',
-    delivery: '廠商配送',
-    shopName: '陽澄湖大閘蟹自營店>',
-  }
-  const wordStyles = {
-    padding: '0 5px',
-    borderRadius: '1px',
-    fontSize: '10px',
-    height: '15px',
-    lineHeight: '15px',
-    color: '#999',
-    backgroundColor: '#f2f2f7',
-    marginRight: '5px',
-  }
-  return (
-    <Card
-      src = { state.src }
-  title = { state.title }
-  price = { state.price }
-  vipPrice = { state.vipPrice }
-  shopDescription = { state.shopDescription }
-  delivery = { state.delivery }
-  shopName = { state.shopName }
-  description = {
-    < div className = "search_prolist_attr"
-    style = {{
-        display: 'inline-flex', 
-        margin:'3px 0 1px', 
-        height:'15px'}}>
-        {['鮮活', '禮盒', '國產'].map((item) => {
-        return (
-            <span style = { wordStyles } className = "word" key = { item } >{ item }</span>)})}
-        </div>} />
-);
-};
-export default App;
-
-```
+<CodeBlock src='h5/demo2.tsx'></CodeBlock>
 
 :::
 
@@ -108,49 +32,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from 'react'
-import { Card, Price, Tag } from '@nutui/nutui-react'
-
-const App = () => {
-  const state = {
-    src: '//img10.360buyimg.com/n2/s240x240_jfs/t1/210890/22/4728/163829/6163a590Eb7c6f4b5/6390526d49791cb9.jpg!q70.jpg',
-    title:
-      '【活蟹】湖塘煙雨 陽澄湖大閘蟹公4.5兩 母3.5兩 4對8只 鮮活生鮮螃蟹現貨水產禮盒海鮮水',
-    price: '388',
-    vipPrice: '378',
-    shopDescription: '自營',
-    delivery: '廠商配送',
-    shopName: '陽澄湖大閘蟹自營店>',
-  }
-  const tagStyles = {
-    display: 'inline-block',
-    verticalAlign: 'middle',
-    marginRight: '5px',
-    marginLeft: '2px',
-    height: '14px',
-  }
-  return (
-    <Card
-      src={state.src}
-      title={state.title}
-      price={state.price}
-      vipPrice={state.vipPrice}
-      shopDescription={state.shopDescription}
-      delivery={state.delivery}
-      shopName={state.shopName}
-      priceTag={
-        <img
-          style={tagStyles}
-          src="https://img11.360buyimg.com/jdphoto/s58x28_jfs/t9451/359/415622649/15318/b0943e5d/59a78495N3bd2a9f8.png"
-          alt=""
-        />
-      }
-     />
-  )
-}
-export default App
-```
+<CodeBlock src='h5/demo3.tsx'></CodeBlock>
 
 :::
 
@@ -158,35 +40,7 @@ export default App
 
 :::demo
 
-```tsx
-import React from "react";
-import { Card, Price, Tag } from '@nutui/nutui-react';
-
-const App = () => {
-  const state = {
-    src:
-      '//img10.360buyimg.com/n2/s240x240_jfs/t1/210890/22/4728/163829/6163a590Eb7c6f4b5/6390526d49791cb9.jpg!q70.jpg',
-    title: '【活蟹】湖塘煙雨 陽澄湖大閘蟹公4.5兩 母3.5兩 4對8只 鮮活生鮮螃蟹現貨水產禮盒海鮮水',
-    price: '388',
-    vipPrice: '378',
-    shopDescription: '自營',
-    delivery: '廠商配送',
-    shopName: '陽澄湖大閘蟹自營店>',
-  }
-  return (
-    <Card
-        src = { state.src }
-        title = { state.title }
-        price = { state.price }
-        vipPrice = { state.vipPrice }
-        shopDescription = { state.shopDescription }
-        delivery = { state.delivery }
-        shopName = { state.shopName }
-        tag = { <div> 自定義店鋪介紹 </div>}> </Card>
-    );
-};
-export default App;
-```
+<CodeBlock src='h5/demo4.tsx'></CodeBlock>
 
 :::
 
@@ -194,36 +48,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from "react";
-import { Card, Price, Tag } from '@nutui/nutui-react';
-
-const App = () => {
-  const state = {
-    src:
-      '//img10.360buyimg.com/n2/s240x240_jfs/t1/210890/22/4728/163829/6163a590Eb7c6f4b5/6390526d49791cb9.jpg!q70.jpg',
-    title: '【活蟹】湖塘煙雨 陽澄湖大閘蟹公4.5兩 母3.5兩 4對8只 鮮活生鮮螃蟹現貨水產禮盒海鮮水',
-    price: '388',
-    vipPrice: '378',
-    shopDescription: '自營',
-    delivery: '廠商配送',
-    shopName: '陽澄湖大閘蟹自營店>',
-  }
-
-  return (
-    <Card
-       src = { state.src }
-        title = { state.title }
-        price = { state.price }
-        vipPrice = { state.vipPrice }
-        shopDescription = { state.shopDescription }
-        delivery = { state.delivery }
-        shopName = { state.shopName }
-        extra = { < div style = {{fontSize: '12px'}}>自定義</div>} />
-    );
-};
-export default App;
-```
+<CodeBlock src='h5/demo5.tsx'></CodeBlock>
 
 :::
